clarify runtime macro
