+++ crates/host-macros/src/runtime/parse.rs
@@ -1,0 +1,144 @@
+//! # Parse
+//! 
+//! Parses the runtime macro token stream input into structured values.
+
+use quote::format_ident;
+use syn::parse::{Parse, ParseStream};
+use syn::punctuated::Punctuated;
+use syn::{Ident, LitBool, Path, Result, Token};
+
+/// Configuration for the runtime macro.
+pub struct Config {
+    pub command: bool,
+    pub hosts: Vec<Host>,
+    pub backends: Vec<Path>,
+}
+
+pub struct Hosts(pub Vec<Host>);
+
+/// Information about a WASI host and its configuration.
+pub struct Host {
+    pub type_: Path,
+    pub backend: Option<Path>,
+}
+
+impl Parse for Config {
+    fn parse(input: ParseStream) -> syn::Result<Self> {
+        let mut command = false;
+        let mut hosts = Hosts(Vec::new());
+
+        let settings;
+        syn::braced!(settings in input);
+        let settings = Punctuated::<Opt, Token![,]>::parse_terminated(&settings)?;
+
+        for setting in settings.into_pairs() {
+            match setting.into_value() {
+                Opt::Command(c) => command = c,
+                Opt::Host(h) => hosts = h,
+            }
+        }
+
+        // deduplicate backends (skipping backend-less hosts)
+        let mut backends: Vec<Path> = vec![];
+        for host in &hosts.0 {
+            let Some(backend) = &host.backend else {
+                continue;
+            };
+            if backends.iter().any(|b| b.get_ident() == backend.get_ident()) {
+                continue;
+            }
+            backends.push(backend.clone());
+        }
+
+        Ok(Self {
+            command,
+            hosts: hosts.0,
+            backends,
+        })
+    }
+}
+
+mod kw {
+    syn::custom_keyword!(command);
+    syn::custom_keyword!(hosts);
+}
+
+#[allow(clippy::large_enum_variant)]
+enum Opt {
+    Command(bool),
+    Host(Hosts),
+}
+
+impl Parse for Opt {
+    fn parse(input: ParseStream) -> Result<Self> {
+        let l = input.lookahead1();
+        if l.peek(kw::command) {
+            input.parse::<kw::command>()?;
+            input.parse::<Token![:]>()?;
+            Ok(Self::Command(input.parse::<LitBool>()?.value))
+        } else if l.peek(kw::hosts) {
+            input.parse::<kw::hosts>()?;
+            input.parse::<Token![:]>()?;
+            let list;
+            syn::braced!(list in input);
+            Ok(Self::Host(list.parse()?))
+        } else {
+            Err(l.error())
+        }
+    }
+}
+
+impl Parse for Hosts {
+    fn parse(input: ParseStream) -> Result<Self> {
+        let hosts = Punctuated::<Host, Token![,]>::parse_terminated(input)?;
+        Ok(Self(hosts.into_iter().collect()))
+    }
+}
+
+impl Parse for Host {
+    fn parse(input: ParseStream) -> syn::Result<Self> {
+        let type_ = input.parse::<Path>()?;
+        // `Host: Backend` (a backend-backed host) or bare `Host` (backend-less).
+        let backend = if input.peek(Token![:]) {
+            input.parse::<Token![:]>()?;
+            Some(input.parse::<Path>()?)
+        } else {
+            None
+        };
+        Ok(Self { type_, backend })
+    }
+}
+
+/// Derives a `snake_case` field name from a backend type's final path segment
+/// (e.g. `HttpDefault` -> `http_default`).
+pub fn field_ident(path: &Path) -> Ident {
+    let Some(segment) = path.segments.last() else {
+        return format_ident!("field");
+    };
+
+    let mut snake = String::new();
+    for ch in segment.ident.to_string().chars() {
+        if ch.is_uppercase() {
+            if !snake.is_empty() {
+                snake.push('_');
+            }
+            snake.extend(ch.to_lowercase());
+        } else {
+            snake.push(ch);
+        }
+    }
+
+    format_ident!("{snake}")
+}
+
+/// Derives a host crate's module ident from a host type's final path segment
+/// (e.g. `WasiHttp` -> `omnia_wasi_http`), matching the host-crate naming the
+/// `StoreContext` derive's `#[wasi(...)]` attribute expects.
+pub fn wasi_ident(path: &Path) -> Ident {
+    let Some(segment) = path.segments.last() else {
+        return format_ident!("wasi");
+    };
+
+    let name = segment.ident.to_string().replace("Wasi", "omnia_wasi_").to_lowercase();
+    format_ident!("{name}")
+}
